requirements.txt
anthropic>=0.18.0
openai>=1.12.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
# requirements.txt
openai>=1.0.0
anthropic>=0.8.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pyyaml>=6.0.0
jupyter>=1.0.0
plotly>=5.0.0
krippendorff>=0.5.0  # For reliability calculations